refactor: folder 구조 수정 및 리팩토링
